Move composite CRDT to core.crdt
